Move conversion exceptions under `error` sub-package
