+++ client/src/main/java/org/spine3/validate/error/IllegalConversionArgumentException.java
@@ -18,22 +18,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.validate;
+package org.spine3.validate.error;
 
 /**
- * Signals that an error has been reached unexpectedly while converting from one type to another.
+ * Thrown to indicate that a method has been passed an illegal or
+ * inappropriate argument during the conversion from one type to another.
  *
  * @author Illia Shepilov
+ * @see ConversionError
  */
-@SuppressWarnings("ExceptionClassNameDoesntEndWithException")
-// It is OK, because it is not {@code Exception} in usual meaning.
-// It is occurred when input value cannot be converted to desirable type
-// and repeated input is required.
-public class ConversionError extends Exception {
+public class IllegalConversionArgumentException extends IllegalArgumentException {
 
     private static final long serialVersionUID = 1L;
 
-    public ConversionError(String message) {
-        super(message);
+    private final ConversionError conversionError;
+
+    public IllegalConversionArgumentException(ConversionError conversionError) {
+        super();
+        this.conversionError = conversionError;
+    }
+
+    public ConversionError getConversionError() {
+        return conversionError;
     }
 }
