+++ client/src/test/java/org/spine3/util/ExceptionsShould.java
@@ -22,7 +22,7 @@ package org.spine3.util;
 
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
-import org.spine3.validate.IllegalConversionArgumentException;
+import org.spine3.validate.error.IllegalConversionArgumentException;
 
 import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.test.Tests.assertHasPrivateParameterlessCtor;
